requirements.txt
pandas==1.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.26.1